requirements.txt
requests
tk
python-dotenv
